backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pandas==2.1.3
numpy==1.24.3
matplotlib==3.8.0
seaborn==0.12.2
scikit-learn==1.3.2
jupyter==1.0.0